requirements.txt
black==21.10b0
click==8.0.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pymongo==3.12.1
PyYAML==6.0
regex==2021.11.2
tomli==1.2.2
typing-extensions==3.10.0.2
